Add files via upload
This project is a Spotify clone created using only HTML and CSS. The goal of this clone was to replicate the basic layout and design of Spotify’s interface, focusing on creating a visually appealing and responsive design. This project does not include any backend or functionality such as music streaming, search, or user authentication, but it showcases the use of HTML and CSS to build a static, mobile-responsive layout.